+++ src/network_event/client_event.rs
@@ -5,9 +5,13 @@ use bevy::{
     prelude::*,
 };
 use bevy_renet::renet::{RenetClient, RenetServer};
-use serde::{de::DeserializeOwned, Serialize};
+use bincode::{DefaultOptions, Options};
+use serde::{
+    de::{DeserializeOwned, DeserializeSeed},
+    Serialize,
+};
 
-use super::EventChannel;
+use super::{BuildEventDeserializer, BuildEventSerializer, EventChannel};
 use crate::{
     client::{ClientState, NetworkEntityMap},
     prelude::NetworkChannels,
@@ -23,6 +27,27 @@ pub trait ClientEventAppExt {
     fn add_mapped_client_event<T: Event + Serialize + DeserializeOwned + Debug + MapEntities>(
         &mut self,
     ) -> &mut Self;
+
+    /// Same as [`Self::add_client_event`], but the event will be serialized/deserialized using `S`/`D`
+    /// with access to [`AppTypeRegistry`].
+    ///
+    /// Needed to send events that contain things like `Box<dyn Reflect>`.
+    fn add_client_reflect_event<T, S, D>(&mut self) -> &mut Self
+    where
+        T: Event + Debug,
+        S: BuildEventSerializer<T> + 'static,
+        D: BuildEventDeserializer + 'static,
+        for<'a> S::EventSerializer<'a>: Serialize,
+        for<'a, 'de> D::EventDeserializer<'a>: DeserializeSeed<'de, Value = T>;
+
+    /// Same as [`Self::add_client_reflect_event`], but additionally maps client entities to server before sending.
+    fn add_mapped_client_reflect_event<T, S, D>(&mut self) -> &mut Self
+    where
+        T: Event + Debug + MapEntities,
+        S: BuildEventSerializer<T> + 'static,
+        D: BuildEventDeserializer + 'static,
+        for<'a> S::EventSerializer<'a>: Serialize,
+        for<'a, 'de> D::EventDeserializer<'a>: DeserializeSeed<'de, Value = T>;
 
     /// Same as [`Self::add_client_event`], but uses specified sending and receiving systems.
     fn add_client_event_with<T: Event + Debug, Marker1, Marker2>(
@@ -43,6 +68,34 @@ impl ClientEventAppExt for App {
         self.add_client_event_with::<T, _, _>(
             mapping_and_sending_system::<T>,
             receiving_system::<T>,
+        )
+    }
+
+    fn add_client_reflect_event<T, S, D>(&mut self) -> &mut Self
+    where
+        T: Event + Debug,
+        S: BuildEventSerializer<T> + 'static,
+        D: BuildEventDeserializer + 'static,
+        for<'a> S::EventSerializer<'a>: Serialize,
+        for<'a, 'de> D::EventDeserializer<'a>: DeserializeSeed<'de, Value = T>,
+    {
+        self.add_client_event_with::<T, _, _>(
+            sending_reflect_system::<T, S>,
+            receiving_reflect_system::<T, D>,
+        )
+    }
+
+    fn add_mapped_client_reflect_event<T, S, D>(&mut self) -> &mut Self
+    where
+        T: Event + Debug + MapEntities,
+        S: BuildEventSerializer<T> + 'static,
+        D: BuildEventDeserializer + 'static,
+        for<'a> S::EventSerializer<'a>: Serialize,
+        for<'a, 'de> D::EventDeserializer<'a>: DeserializeSeed<'de, Value = T>,
+    {
+        self.add_client_event_with::<T, _, _>(
+            mapping_and_sending_reflect_system::<T, S>,
+            receiving_reflect_system::<T, D>,
         )
     }
 
@@ -96,6 +149,52 @@ fn mapping_and_sending_system<T: Event + MapEntities + Serialize + Debug>(
     }
 }
 
+fn sending_reflect_system<T, S>(
+    mut events: EventReader<T>,
+    mut client: ResMut<RenetClient>,
+    channel: Res<EventChannel<T>>,
+    registry: Res<AppTypeRegistry>,
+) where
+    T: Event + Debug,
+    S: BuildEventSerializer<T>,
+    for<'a> S::EventSerializer<'a>: Serialize,
+{
+    let registry = registry.read();
+    for event in &mut events {
+        let serializer = S::new(&registry, event);
+        let message =
+            bincode::serialize(&serializer).expect("client reflect event should be serializable");
+        client.send_message(channel.id, message);
+        debug!("sent client reflect event {event:?}");
+    }
+}
+
+fn mapping_and_sending_reflect_system<T, S>(
+    mut events: ResMut<Events<T>>,
+    mut client: ResMut<RenetClient>,
+    entity_map: Res<NetworkEntityMap>,
+    channel: Res<EventChannel<T>>,
+    registry: Res<AppTypeRegistry>,
+) where
+    T: Event + MapEntities + Debug,
+    S: BuildEventSerializer<T>,
+    for<'a> S::EventSerializer<'a>: Serialize,
+{
+    let registry = registry.read();
+    for mut event in events.drain() {
+        event
+            .map_entities(entity_map.to_server())
+            .unwrap_or_else(|e| {
+                panic!("client reflect event {event:?} should map its entities: {e}")
+            });
+        let serializer = S::new(&registry, &event);
+        let message = bincode::serialize(&serializer)
+            .expect("mapped client reflect event should be serializable");
+        client.send_message(channel.id, message);
+        debug!("sent mapped client reflect event {event:?}");
+    }
+}
+
 /// Transforms [`T`] events into [`FromClient<T>`] events to "emulate"
 /// message sending for offline mode or when server is also a player
 fn local_resending_system<T: Event + Debug>(
@@ -129,6 +228,38 @@ fn receiving_system<T: Event + DeserializeOwned + Debug>(
     }
 }
 
+fn receiving_reflect_system<T, D>(
+    mut client_events: EventWriter<FromClient<T>>,
+    mut server: ResMut<RenetServer>,
+    channel: Res<EventChannel<T>>,
+    registry: Res<AppTypeRegistry>,
+) where
+    T: Event + Debug,
+    D: BuildEventDeserializer,
+    for<'a, 'de> D::EventDeserializer<'a>: DeserializeSeed<'de, Value = T>,
+{
+    let registry = registry.read();
+    for client_id in server.clients_id() {
+        while let Some(message) = server.receive_message(client_id, channel.id) {
+            // Set options to match `bincode::serialize`.
+            // https://docs.rs/bincode/latest/bincode/config/index.html#options-struct-vs-bincode-functions
+            let options = DefaultOptions::new()
+                .with_fixint_encoding()
+                .allow_trailing_bytes();
+            let mut deserializer = bincode::Deserializer::from_slice(&message, options);
+            match D::new(&registry).deserialize(&mut deserializer) {
+                Ok(event) => {
+                    debug!("received reflect event {event:?} from client {client_id}");
+                    client_events.send(FromClient { client_id, event });
+                }
+                Err(e) => {
+                    error!("unable to deserialize reflect event from client {client_id}: {e}")
+                }
+            }
+        }
+    }
+}
+
 /// An event indicating that a message from client was received.
 /// Emited only on server.
 #[derive(Clone, Copy)]
@@ -139,14 +270,16 @@ pub struct FromClient<T> {
 
 #[cfg(test)]
 mod tests {
-    use bevy::ecs::{
-        entity::{EntityMap, MapEntitiesError},
-        event::Events,
-    };
-    use serde::Deserialize;
+    use bevy::ecs::event::Events;
 
     use super::*;
-    use crate::{test_network::TestNetworkPlugin, ReplicationPlugins};
+    use crate::{
+        network_event::test_events::{
+            DummyEvent, ReflectEvent, ReflectEventDeserializer, ReflectEventSerializer,
+        },
+        test_network::TestNetworkPlugin,
+        ReplicationPlugins,
+    };
 
     #[test]
     fn sending_receiving() {
@@ -196,6 +329,66 @@ mod tests {
     }
 
     #[test]
+    fn sending_receiving_reflect() {
+        let mut app = App::new();
+        app.add_plugins(ReplicationPlugins)
+            .register_type::<Transform>()
+            .add_client_reflect_event::<ReflectEvent, ReflectEventSerializer, ReflectEventDeserializer>()
+            .add_plugin(TestNetworkPlugin);
+
+        app.world
+            .resource_mut::<Events<ReflectEvent>>()
+            .send(ReflectEvent {
+                entity: Entity::PLACEHOLDER,
+                component: Transform::IDENTITY.clone_value(),
+            });
+
+        app.update();
+        app.update();
+
+        let transforms: Vec<_> = app
+            .world
+            .resource_mut::<Events<FromClient<ReflectEvent>>>()
+            .drain()
+            .map(|event| Transform::from_reflect(&*event.event.component).unwrap())
+            .collect();
+        assert_eq!(transforms, [Transform::IDENTITY]);
+    }
+
+    #[test]
+    fn mapping_and_sending_receiving_reflect() {
+        let mut app = App::new();
+        app.add_plugins(ReplicationPlugins)
+            .register_type::<Transform>()
+            .add_mapped_client_reflect_event::<ReflectEvent, ReflectEventSerializer, ReflectEventDeserializer>()
+            .add_plugin(TestNetworkPlugin);
+
+        let client_entity = Entity::from_raw(0);
+        let server_entity = Entity::from_raw(client_entity.index() + 1);
+        app.world
+            .resource_mut::<NetworkEntityMap>()
+            .insert(server_entity, client_entity);
+
+        app.world
+            .resource_mut::<Events<ReflectEvent>>()
+            .send(ReflectEvent {
+                entity: client_entity,
+                component: Transform::IDENTITY.clone_value(),
+            });
+
+        app.update();
+        app.update();
+
+        let mapped_entities: Vec<_> = app
+            .world
+            .resource_mut::<Events<FromClient<ReflectEvent>>>()
+            .drain()
+            .map(|event| event.event.entity)
+            .collect();
+        assert_eq!(mapped_entities, [server_entity]);
+    }
+
+    #[test]
     fn local_resending() {
         let mut app = App::new();
         app.add_plugins(ReplicationPlugins)
@@ -212,15 +405,5 @@ mod tests {
 
         let client_events = app.world.resource::<Events<FromClient<DummyEvent>>>();
         assert_eq!(client_events.len(), 1);
-    }
-
-    #[derive(Deserialize, Serialize, Debug)]
-    struct DummyEvent(Entity);
-
-    impl MapEntities for DummyEvent {
-        fn map_entities(&mut self, entity_map: &EntityMap) -> Result<(), MapEntitiesError> {
-            self.0 = entity_map.get(self.0)?;
-            Ok(())
-        }
     }
 }